Event refactor
All events are now subscribed to in OnEnable and unsubscribed from in OnDisable

Note: Singletons that assign themselves as Instance in Awake and have events have had to be placed first in script execution order, in the appropriate order.

For now, the order is PlayerController > ScoreManager > CameraController.

Any new singletons with events will have to be placed in this order appropriately.